Set filter-FORWARD drop on restart after enabling IP forwarding
The daemon sets the host's filter-FORWARD firewall chain policy to
DROP if it enables IP forwarding.

With iptables, that policy is preserved over a daemon restart (the
chain is not deleted, it's a built-in chain). But, with nftables,
the daemon's table is re-created on restart, so the filter-FORWARD
drop policy is lost.

So, when setting the policy to "drop" after enabling IP forwarding,
record that in the datastore. Every subsequent daemon restart will
then set the policy to drop (for iptables or nftables).

To stop the daemon from setting the policy, it must be started with
option '--ip-forward-no-drop'.

Signed-off-by: Rob Murray <rob.murray@docker.com>