Migrate iptables filter-FORWARD policy on upgrade
When a pre-29.0 daemon has been running, and it set the iptables
filter-FORWARD policy to DROP because it enabled IP forwarding, if
the daemon is stopped, upgraded to >=29.0 and nftables is enabled...

The iptables policy will still be DROP. The newly started daemon
won't need to enable IP forwarding, and there won't be a record in
the datastore indicating that it needs to be set.

So, when migrating from iptables to nftables, check the iptables
filter-FORWARD policy. If it's DROP, set the nftables policy to
drop, and note that in the datastore for future restarts. Then,
set the iptables policy to ACCEPT, so that it doesn't drop packets
for published ports (because the allow rules for them will only
be in the nftables chain).

Signed-off-by: Rob Murray <rob.murray@docker.com>