contrib: remove syntax (Dockerfile syntax) files
The Dockerfile syntax is now maintained as part of BuildKit in the
https://github.com/moby/buildkit repository, but our contrib still
had syntax highlighting files for various editors;

- The TextMate (`.tmbundle`) syntax was last update 9 months ago
  in c2029cb2574647e4bc28ed58486b8e85883eedb9, but this was a stale
  pull request from 2018, so at least 7 Years behind on updates to
  the syntax.
- The vim syntax is now integrated in upstream, and our copy was
  remove 6 Years ago in 5511f45767e78b136c292dbb989296fb6e4d12fc.
- The Nano syntax was added 10 Years ago in 76bc44fb912c66498b8e6e714ce4aabd9eaef8dc,
  and received an update 9 Years ago in 4cb71f80823af345d063cf0ad657e73ce9caa75f,
  and received no contributions since.
- The kate syntax was already unmaintained, and removed 9 Years ago
  in a2f695d7fae47481eba68b92b33257605e4cbada.

I think it's fair to state that these are unmaintained and severely outdated.
I _do_ think it would be good to have maintained versions of such files,
as it can improve the experience for users. However, this would be more
suitable for the BuildKit project, and I'm not sure if the current copies
of these files are a good starting point (if needed, they can still be
obtained from git history).

This patch removes the files.

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>