add `go.work` instead of `replace` directives for development
Using a `go.work` file lets us use the local (unreleased) `api` and
`client` modules during development without having to add `replace`
directives in `go.mod` (which have to be removed for releases).

Unlike `replace` rules, `go.work` is not transitive, so not visible
to users of the module (which *will* mean that unreleased versions
of the client or "v2" module may be non-functional if they depend
on changes in the api or client that are not yet release, and not
yet reflected in their `go.mod`).

Use `GOWORK=off` to verify what "external" users see.

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>