vendor what's in go.mod?
This makes the vendor directory reflect what's specified in go.mod;
downside is that this may make the code non-buildable without go.work
enabled (as the vendor directory doesn't reflect what go.work uses),
but it does reduce code-churn w.r.t. vendored files being removed
while go.work is present.

Possibly we can make vendor handle generating a pseudo-version and
update the vendor "as usual" without using a replace rule, but
that needs to be looked into if there's an option for that.

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>