daemon/config: disable the userland proxy by default
The userland proxy is currently used to handle connections to mapped
ports on the loopback address, or from other containers connected to the
same network, when they target a host's IP address (e.g. their bridge IP
address, or one of the host's 'external' IP addresses).

When the userland proxy is disabled, connections coming from other
containers can be DNATed as usual — but the br_netfilter module is
required. For loopback connections, IPv4 packets can be DNATed too, but
not IPv6 packets (so, it's not possible to map a port to ::1).

As we discovered, there are some special setups where br_netfilter might
not be available — commit e7bd60ee added an escape hatch to ignore
module loading errors. (Before commit 5c499fc4, br_netfilter was loaded
unconditionally, but no loading errors were reported.)

Regarding ::1 mappings, Linux provides no way to disable IPv4 globally.
So there are little chances that ::1 is available, but not 127.0.0.1.
Mapping a different port to both loopback addresses would be silly.
Thus, there are probably no use cases where 127.0.0.1 couldn't be used
in place of ::1. But re-enabling the userland proxy is still an option
if needed.

The userland proxy also has some drawbacks: it consumes unnecessary
resources, and mangles the connection's source address, which can
confuse users when they look at their application logs.

Signed-off-by: Albin Kerouanton <albinker@gmail.com>