api/types/container: inline PortSet, PortMap in types
Use the underlying definition in types, instead of the `PortSet` and `PortMap`
types as intermediates.

This also shows that there's ambiguity in these definitions, because `nat.Port`
(`container.PortRangeProto`) allows either a individual port ("8080/tcp")
or a range of ports ("8080-8090/tcp"). The latter unlikely is supported
by our code.

That behavior was introduced in [moby@47272f9], which changed `nat.Port` to
either be a Port, or a range of ports. However, it's debatable if that should
have modified the existing type or introduced a new type that's only used
for places where a range is expected.

To reduce ambiguity, we need to evaluate where the `nat.Port` / `PortRangeProto`
type is used, and if there's any places where it SHOULD accept a range.
For other places, we should consider having an alias with a more suitable
name (`PortProto`) - at least to be clearer on intent.

[moby@47272f9]: https://github.com/moby/moby/commit/47272f9cc563ea90d1f86df044f5429d15a37e4f

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>