inte/network/bridge/iptablesdoc: re-generate golden files
Commit da8ecc62dd765b15df84c3aa6b83dcb7a81d4ffa from repo git.netfilter.org/iptables
changed the output format of `iptables --numeric --list` (see [1]) —
proto are represented numerically. This commit was shipped in Debian 12.
That's the format used to generate iptablesdoc golden files initially.

But subsequently, commit 34f085b1607364f4eaded1140060dcaf965a2649
reverted that change to represent proto as strings, even when
`--numeric` is specified (see [2]). This was done to address a bug
report (see [3]). Debian 13 ships that commit.

So, re-generate golden files to match the new output format.

[1]: https://git.netfilter.org/iptables/commit/?id=da8ecc62dd765b15df84c3aa6b83dcb7a81d4ffa
[2]: https://git.netfilter.org/iptables/commit/?id=34f085b1607364f4eaded1140060dcaf965a2649
[3]: https://bugzilla.netfilter.org/show_bug.cgi?id=1729

Signed-off-by: Albin Kerouanton <albinker@gmail.com>