pkg/jsonmessage: JSONMessage: deprecate From, Time, and TimeNano fields
On API version v1.22 and older, the `JSONMessage` was used to produce the
`/events` response. However, commit 72f1881df102fce9ad31e98045b91c204dd44513
introduced an `events.Message` type that replaced the use of `JSONMessage`
for that purpose.

The `JSONMessage` is no longer used to unmarshal these messages, and users
of this package should not depend on these fields, and use the `events.Message`
type instead for unmarshaling the `/events` response.

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>