api/types/network: generate network summary struct
Generate the "api/types/network".Summary struct type from the Swagger
definition. Switch around the aliasing such that the Inspect type
aliases the Summary type. The Inspect struct would be a superset
of the Summary struct when the time comes to return more details when
inspecting a network versus listing networks.

Signed-off-by: Cory Snider <csnider@mirantis.com>