d/network: filter networks individually
Internally a network is represented by either a libnetwork.Network or a
swarmapi.Network. The daemon functions backing the API server map
these values to the Engine API network.Inspect type on demand. Since
they have to convert, the functions to get a list of networks have to
loop over the slice of Networks and append them to a slice of
network.Inspect values.

The function used to filter the list of networks by a user-supplied
predicate takes a []network.Inspect and returns a shorter slice.
Therefore the daemon functions backing the API server have to loop
through the list twice; once to convert, and again inside the
FilterNetworks function to prune networks from the slice. Each time an
item is deleted from a slice, all items at higher indices need to be
copied to lower indices in the backing array.

Replace FilterNetworks with a function that accepts a single network
summary and returns a boolean. The daemon network-list functions can
filter in the conversion loop, which cuts down on a nontrivial amount of
copying. Split the validation of the filter predicate from filter
evaluation to both make it more ergonomic to use inside loops, and to
make invalid states (a filter with an ill-formed predicate)
unrepresentable.

This change also future-proofs the network filter for when we change
network.Inspect to a distinct type that embeds a network.Summary.
Callers that need to build a filtered list of network.Inspect values
will be able to evaluate the filter on the embedded Summary then append
the full Inspect value of matching networks to the output slice.

Signed-off-by: Cory Snider <csnider@mirantis.com>