d/network: fix filtering networks by ID or Name
The network filter has an hidden 'idOrName' term. API clients cannot use
this term as it will not pass ValidateFilters: it is not listed in the
acceptedFilters set. And it's a good thing, too: it is completely broken
and not fit for purpose. When a filter contains an idOrName term, the
term values are ignored and instead the filter tests whether the 'id' or
'name' terms match the Name field of the network. Unless the filter
contains both 'id' and 'name' terms, the match will evaluate to true!
None of the daemon-internal users of 'idOrName' set either term,
therefore it has the same effect as if the filter did not contain the
term in the first place.

Filtering networks by id-or-name is a quirky thing that the daemon needs
to do to uphold its end of the Engine API contract, not something that
would be of use to clients. Instead of fixing up the idOrName filter,
add an exported field to the Filter struct to enable the internal-only
behaviour of having the 'id' term match on either the network Name or
ID.

Signed-off-by: Cory Snider <csnider@mirantis.com>