daemon: validate filter terms in network.NewFilter
Filter-term validation does not belong in the API module. Clients should
not be making any assumptions about which terms the daemon understands.
Users should not need to upgrade their clients to use filter terms
introduced in a newer daemon. Move the network filter validation from
the api module into the daemon.

Have network.NewFilter validate the filter terms, enforcing the
invariant that any network.Filter is a well-formed filter for networks.
Lift the conversion from filters.Args to network.Filter into the API
endpoint handlers so that validation errors continue to be returned
separately from errors that occur when listing networks. As a side
benefit, this also affords the API handlers configuration of the filter
options (IDAlsoMatchesName) without additional plumbing.

Unify the filter terms available when listing networks and pruning
networks. Support filtering on the absence of a label or network
creation time when listing networks. And support all the network listing
filter terms when pruning networks.

Signed-off-by: Cory Snider <csnider@mirantis.com>