libcontainerd/remote: task.Exec: remove redundant closing of io
The `rio` var is not set until `task.Exec` returns. If it returns with
a non-nil error, [task.Exec] itself already handles closing the io before
returning (either directly or in a defer), so we don't have to set up a
defer on error. After successfully executing `task.Exec`, both [task.Start]
and [process.Start] already have a defer to close on error, making the
defer redundant.

[task.Exec]: https://github.com/containerd/containerd/blob/v2.1.4/client/task.go#L424-L468
[task.Start]: https://github.com/containerd/containerd/blob/v2.1.4/client/task.go#L243-L261
[process.Start]: https://github.com/containerd/containerd/blob/v2.1.4/client/process.go#L123-L144

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>